+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - Better locations error handling
 - Improved parsing of 853/863 fields
+- Updated nypl_core_util and loaded deployment helper class from there
 
 ## 0.0.1 - 2020-09-17
 ### Added