+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 0.0.2 - Unreleased
+### Added
+- Added rubocop for linting and linted all files
+### Fixed
+- Better locations error handling
+- Improved parsing of 853/863 fields
+- Updated nypl_core_util and loaded deployment helper class from there
+
 ## 0.0.1 - 2020-09-17
 ### Added
 - Initial commit containing basic structure and location parser